Fat Refactoring Organization
This is a big refactoring for reorganizing the project and the import in
a more pythonic way. We will always have to execute the main file fdb.py
if you do not want import fails.
If you really need to work on a particular file, change temporarly its
imports to execute from this file.